Fix conffile upgrade logic when upgrading from wheezy
When upgrading from wheezy with apache2.2-common but without apache2
installed to jessie, part of the conffile handling logic would not run
because the one instance of the MOVED_CONFFILES_IN variable has been
overlooked when refactoring the logic to use the MOVED_CONFFILES
variable with different contents.

This causes the content from wheezy to be kept for all renamed
conffiles. Fix this so that at least upgrades from wheezy directly to
the newest jessie point release won't have this problem.

Installations that have already been upgraded to a broken version will
be fixed during the upgrade to stretch.

See bug #794933