fix: mislead safety precautions
These safety measures are meant well but are generally in practice not a
good idea. Including a config file is supposed to mean that one wants
the contained directives to have effect.
Removing the enclosing <IfModule> block means that the directives
contained will either have effect or cause an error instead of not
having effect AND not causing an error.
This seems a lot better than them being silently ignored because the
module they are for is not loaded which may be a mistake.
This effectively removes the second layer of "enabling directives" which
has shown to be confusing for users not least because of having the
undesirable effect of included directives being ignored.
The first layer, linking the .load and .conf files to the corresponding
*-enabled directory, is not affected by this change, is much more
straightforward to understand on its own, and should be the preferred
way of disabling directives.