Fix: confusing and impractical naming
This makes it more likely that enabling this vhost makes it the first
and thus the default vhost.
A lot of users have run into this confusion because they don’t
understand that _default_ is not what makes a vhost the default.
They get unintended results when they suddenly add a vhost that starts
with a character that is alphabetically sorted before the 'd' in
'default'.